+++ src/lib/components/nft/NftCard.tsx
@@ -1,14 +1,13 @@
 import type { Addr, HexAddr32, Nullable, Option } from "lib/types";
 
-import { Box, Flex, Image, Text } from "@chakra-ui/react";
+import { Box, Flex, Text } from "@chakra-ui/react";
 import { AmpEvent, track } from "lib/amplitude";
 import { useEvmConfig, useMoveConfig } from "lib/app-provider";
-import { NFT_IMAGE_PLACEHOLDER } from "lib/data";
 import { useFormatAddresses } from "lib/hooks/useFormatAddresses";
 import { useMetadata } from "lib/services/nft";
-import { getIpfsUrl } from "lib/services/utils";
 
 import { AppLink } from "../AppLink";
+import { NftImage } from "./NftImage";
 
 interface NftCardProps {
   collectionAddress: Addr;
@@ -45,17 +44,15 @@ export const NftCard = ({
         onClick={() => track(AmpEvent.USE_NFT_CARD, { showCollection })}
       >
         <Box mb={2} paddingBottom="100%" position="relative" width="100%">
-          <Image
+          <NftImage
             background="gray.900"
             backgroundPosition="center"
             borderRadius="8px"
-            fallbackSrc={NFT_IMAGE_PLACEHOLDER}
-            fallbackStrategy="beforeLoadOrError"
             height="100%"
+            imageUrl={metadata?.image}
             left={0}
             objectFit="contain"
             position="absolute"
-            src={metadata?.image ? getIpfsUrl(metadata.image) : undefined}
             top={0}
             width="100%"
           />